+++ NEWS.md
@@ -1,3 +1,11 @@
+# RBERT 0.1.11
+
+* Added parameter to shush verbose `extract_features`.
+* Removed vestigal `use_one_hot_embeddings` parameter from everywhere.
+* `extract_features` now can take plain character vectors as input.
+* `extract_features` now can take a single checkpoint directory or model name,
+rather than three separate file paths.
+
 # RBERT 0.1.7
 
 * Updated `extract_features` to return tidy tibbles (@jonthegeek, #29)